Added: SevenZipExtractor.ExtractFile with streaming , SevenZipCompressor.CompressStream.
Fixed: throwing an Exception while extracting and a file already exists (id# 2224), UnpackedSize throws an Exception (id# 2224), ExtractFiles did not work properly at all, extraction progress sometimes was wrong, really a lot of other fixes which improved the stability.

Unfortunatetly, WebResponseStream is not supported, it will be in the 0.32 release.

git-svn-id: https://sevenzipsharp.svn.codeplex.com/svn@18703 9b283d60-5439-405e-af05-b73fd8c4d996